+++ java/lance-jni/src/blocking_dataset.rs
@@ -2471,7 +2471,7 @@ fn inner_get_indexes<'local>(
     let array_list = env.new_object("java/util/ArrayList", "()V", &[])?;
 
     for index_meta in indexes.iter() {
-        let java_index = create_java_index(env, index_meta)?;
+        let java_index = index_meta.into_java(env)?;
         env.call_method(
             &array_list,
             "add",
@@ -2481,192 +2481,6 @@ fn inner_get_indexes<'local>(
     }
 
     Ok(array_list)
-}
-
-fn create_java_index<'local>(
-    env: &mut JNIEnv<'local>,
-    index_meta: &IndexMetadata,
-) -> Result<JObject<'local>> {
-    // Create UUID
-    let uuid_str = index_meta.uuid.to_string();
-    let uuid_obj = env
-        .call_static_method(
-            "java/util/UUID",
-            "fromString",
-            "(Ljava/lang/String;)Ljava/util/UUID;",
-            &[JValue::Object(&env.new_string(&uuid_str)?.into())],
-        )?
-        .l()?;
-
-    // Create fields list (List<Integer>)
-    let fields_list = env.new_object("java/util/ArrayList", "()V", &[])?;
-    for field in &index_meta.fields {
-        let int_obj = env
-            .call_static_method(
-                "java/lang/Integer",
-                "valueOf",
-                "(I)Ljava/lang/Integer;",
-                &[JValue::Int(*field)],
-            )?
-            .l()?;
-        env.call_method(
-            &fields_list,
-            "add",
-            "(Ljava/lang/Object;)Z",
-            &[JValue::Object(&int_obj)],
-        )?;
-    }
-
-    // Create name
-    let name = env.new_string(&index_meta.name)?;
-
-    // Create fragments list (List<Integer>) - can be null
-    let fragments_list = if let Some(bitmap) = &index_meta.fragment_bitmap {
-        let list = env.new_object("java/util/ArrayList", "()V", &[])?;
-        for frag_id in bitmap.iter() {
-            let int_obj = env
-                .call_static_method(
-                    "java/lang/Integer",
-                    "valueOf",
-                    "(I)Ljava/lang/Integer;",
-                    &[JValue::Int(frag_id as i32)],
-                )?
-                .l()?;
-            env.call_method(
-                &list,
-                "add",
-                "(Ljava/lang/Object;)Z",
-                &[JValue::Object(&int_obj)],
-            )?;
-        }
-        list
-    } else {
-        JObject::null()
-    };
-
-    // Create indexDetails byte array - can be null
-    let index_details_bytes = if let Some(details) = &index_meta.index_details {
-        let bytes = prost::Message::encode_to_vec(details.as_ref());
-        let byte_array = env.byte_array_from_slice(&bytes)?;
-        byte_array.into()
-    } else {
-        JObject::null()
-    };
-
-    // Create createdAt Instant - can be null
-    let created_at = if let Some(dt) = index_meta.created_at {
-        let millis = dt.timestamp_millis();
-        env.call_static_method(
-            "java/time/Instant",
-            "ofEpochMilli",
-            "(J)Ljava/time/Instant;",
-            &[JValue::Long(millis)],
-        )?
-        .l()?
-    } else {
-        JObject::null()
-    };
-
-    // Create baseId Integer - can be null
-    let base_id = if let Some(id) = index_meta.base_id {
-        env.call_static_method(
-            "java/lang/Integer",
-            "valueOf",
-            "(I)Ljava/lang/Integer;",
-            &[JValue::Int(id as i32)],
-        )?
-        .l()?
-    } else {
-        JObject::null()
-    };
-
-    // Determine index type from index_details type_url
-    let index_type = determine_index_type(env, &index_meta.index_details)?;
-
-    // Call Index.create() static method
-    let index_obj = env.call_static_method(
-        "org/lance/index/Index",
-        "create",
-        "(Ljava/util/UUID;Ljava/util/List;Ljava/lang/String;JLjava/util/List;[BILjava/time/Instant;Ljava/lang/Integer;Lorg/lance/index/IndexType;)Lorg/lance/index/Index;",
-        &[
-            JValue::Object(&uuid_obj),
-            JValue::Object(&fields_list),
-            JValue::Object(&name.into()),
-            JValue::Long(index_meta.dataset_version as i64),
-            JValue::Object(&fragments_list),
-            JValue::Object(&index_details_bytes),
-            JValue::Int(index_meta.index_version),
-            JValue::Object(&created_at),
-            JValue::Object(&base_id),
-            JValue::Object(&index_type),
-        ],
-    )?.l()?;
-
-    Ok(index_obj)
-}
-
-fn determine_index_type<'local>(
-    env: &mut JNIEnv<'local>,
-    index_details: &Option<Arc<prost_types::Any>>,
-) -> Result<JObject<'local>> {
-    let type_name = if let Some(details) = index_details {
-        // Extract type name from type_url (e.g., ".lance.index.BTreeIndexDetails" -> "BTREE")
-        let type_url = &details.type_url;
-        let type_part = type_url.split('.').next_back().unwrap_or("");
-        let lower = type_part.to_lowercase();
-
-        if lower.contains("btree") {
-            Some("BTREE")
-        } else if lower.contains("bitmap") {
-            Some("BITMAP")
-        } else if lower.contains("labellist") {
-            Some("LABEL_LIST")
-        } else if lower.contains("inverted") {
-            Some("INVERTED")
-        } else if lower.contains("ngram") {
-            Some("NGRAM")
-        } else if lower.contains("zonemap") {
-            Some("ZONEMAP")
-        } else if lower.contains("bloomfilter") {
-            Some("BLOOM_FILTER")
-        } else if lower.contains("ivfhnsw") {
-            if lower.contains("sq") {
-                Some("IVF_HNSW_SQ")
-            } else if lower.contains("pq") {
-                Some("IVF_HNSW_PQ")
-            } else {
-                Some("IVF_HNSW_FLAT")
-            }
-        } else if lower.contains("ivf") {
-            if lower.contains("sq") {
-                Some("IVF_SQ")
-            } else if lower.contains("pq") {
-                Some("IVF_PQ")
-            } else {
-                Some("IVF_FLAT")
-            }
-        } else if lower.contains("vector") {
-            Some("VECTOR")
-        } else {
-            None
-        }
-    } else {
-        None
-    };
-
-    match type_name {
-        Some(name) => {
-            let index_type = env
-                .get_static_field(
-                    "org/lance/index/IndexType",
-                    name,
-                    "Lorg/lance/index/IndexType;",
-                )?
-                .l()?;
-            Ok(index_type)
-        }
-        None => Ok(JObject::null()),
-    }
 }
 
 #[no_mangle]
